Add envoy, and support for docker proxies

+++ scripts/test.sh
@@ -4,7 +4,7 @@ here=$(dirname $0)
 base=$here/..
 results=$base/results
 
-nginx_docker_id=$(docker run --rm -d -p 80:80 nginx)
+nginx_docker_id=$(docker run --rm -d --network host -p 80:80 --name nginx nginx)
 read -r -d '' plotscript <<'EOF'
 set terminal png
 set xlabel "distribution"
@@ -21,15 +21,22 @@ function record_server() {
     cmd="$3"
     port=$4
     mkdir -p $results
-    (cd $dir; exec $cmd) &
+    (cd $dir; exec $(eval echo $cmd) 2>&1 > /dev/null) &
     cmd_pid=$!
     echo pid is $cmd_pid
     until nc -z localhost $port; do
         sleep 1;
     done
     echo $kind is up, starting the test
-    ~/work/wrk2/wrk -t2 -c100 -d30s -R2000 --latency http://localhost:$port/index.html | grep -A8 HdrHistogram | tail -8 | sed -re 's/us$/E-6/g;s/ms$/E-3/g;s/\.?\0+%/%/g' > $results/$kind
-    kill $cmd_pid
+    ~/work/wrk2/wrk -t2 -c100 -d240s -R2000 --latency http://localhost:$port/index.html | tee $results/$kind.full | grep -A8 HdrHistogram | tail -8 | sed -re 's/us$/E-6/g;s/ms$/E-3/g;s/\.?\0+%/%/g' > $results/$kind
+    if [ -z "${stop_cmd}" ]; then
+        kill $cmd_pid;
+    else
+        echo $stop_cmd
+        $stop_cmd;
+        unset stop_cmd;
+        echo $stop_cmd
+    fi
     plotscript="$plotscript \"$kind\" using 0:2:xticlabel(1) title \"$kind\" with lines,"
 }
 
@@ -40,5 +47,5 @@ for profile in $(find $base -name run.profile); do
     kind=$(basename $dir);
     record_server "$dir" "$kind" "$cmd" $port;
 done
-(cd $results; echo "$plotscript" | gnuplot > plot.png)
-docker stop $nginx_docker_id
+(cd $results; echo "$plotscript" | gnuplot > plot_$(date -Iminutes).png)
+docker stop nginx
